theme context applied in a new branch themeContext
