+++ product-filter/src/index.js
@@ -2,12 +2,18 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import { ThemeContextProvider } from './context/themeContext';
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
+   
+   <ThemeContextProvider>
+
     <App />
+
+   </ThemeContextProvider>
   </React.StrictMode>
 );
